requirements.txt
numpy==1.26.4
h5py
pyyaml
pandas
scikit-learn
-f https://download.pytorch.org/whl/cu116/torch_stable.html
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1+cu116
-f https://data.pyg.org/whl/torch-1.13.1+cu116.html
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-geometric
opencv-python
tensorflow